pyproject.toml
[project]
name = "langmem-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-deepseek>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph-checkpoint-redis>=0.1.2",
    "langmem>=0.0.30",
    "numpy>=2.3.4",
    "python-dotenv>=1.2.1",
]
